+++ .github/actions/validate_module/entrypoint.sh
@@ -1,0 +1,3 @@
+#! /usr/bin/env bash
+pdk convert --force
+pdk validate